Refactor circe package
